Add Thumbprint support for opaque signers
This commit allows getting a thumbprint for the type OpaqueSigner. So if
a key is backed by a custom signer, for example, a YubiKey or a cloud
signer, we should be able to get the thumbprint for it as long as the
public key is one of the supported ones.

Signed-off-by: Mariano Cano <mariano.cano@gmail.com>