Require specifying algorithms for ParseEncrypted (#69)
Also for jwt.ParseEncryptedAndSigned.

This allows specifying multiple accepted algorithms for both the key
encryption algorithm and the content encryption algorithm, with a
semantic of "accept any of these." It does not support a blanket "accept
all" setting. The user must specify an explicit list of acceptable
algorithms.

This is a followup to #64. If we like this approach I'll prepare an
additional followup that adds similar parameters to ParseSigned.

Right now this is an API breaking change, and the plan is to move to v4.
Alternately, we could specify these as two new functions and deprecate
the old ones, then later down the road do the API-breaking revision by
deleting the deprecated functions.