Revert #130: JSONWebKeySet: ignore unsupported key types (#137)
As discussed in #136, this was a breaking change. Implementing
`UnmarshalJSON` on `JSONWebKeySet` broke unmarshaling into structs that
embed `JSONWebKeySet` and also include other fields (and that do not
have `UnmarshalJSON` implementations themselves). Before #130, those
fields would get unmarshaled using the default behavior for
`encoding/json`. After #130, the custom `UnmarshalJSON` would get called
for the whole parent struct, but because its receiver is the embedded
`JSONWebKeySet` it would not be able to set any of the sibling fields.

Revert for now to undo the breakage. We should find another way to solve
the problem of accepting unrecognized keys.

Note: this also reverts #135 which landed after #130.