Reject JWS with an unprotected critical b64 header
The only critical header that go-jose supports is b64 from RFC7797.

go-jose correctly only respects that header if it appears in a protected
header. go-jose correctly rejects unknown critical headers.

However, go-jose does not reject a JWS which contains a critical b64
unprotected header.

I don't believe this has any security impact, as the only place this is exposed
is via the Signature header members which expose unprotected headers, and so a
library user is already aware they are not to be trusted. The 'critical'
behavior (of not base64-encoding the value, defined in RFC7797) is not
influenced here.