jws: RFC7797 unencoded payload serialization
The JOSE RFC7797 “b64” header parameter was already respected during
signature computation, but serialization always base64url-encoded the payload,
even when "b64" was set to false. This produced valid RFC 7797 signatures, but
invalid payloads (i.e., base64-encoded payloads instead of unencoded
payloads).

This commit changes the Payload data type in the rawJSONWebSignature to
string and delays its base64-decoding to the sanitized() function to
be able to check the b64 header. It further updates CompactSerialize()
and FullSerialize() to correctly emit the unencoded payload in
case "64" is false as required by RFC7797. For all other cases, the
payload continues to be base64-encoded as before.